notebooks/requirements.txt
numpy==1.26.4
pandas==2.2.2
torch==2.3.0
sentence-transformers==2.7.0
scikit-learn==1.4.2
jupyterlab==4.2.1
seaborn==0.12.2
datasets==2.19.1